Delete Runners::HardcodedRunner
