--- lib/retest/runners/hardcoded_runner.rb
@@ -1,9 +1,0 @@
-module Retest
-  module Runners
-    class HardcodedRunner < Runner
-      def run(_ = nil)
-        system command
-      end
-    end
-  end
-end
