+++ test/retest/runners/runner_test.rb
@@ -5,10 +5,20 @@ module Retest
   module Runners
     class RunnerTest < MiniTest::Test
       def setup
-        @subject = Runner.new("echo 'hell world'")
+        @subject = Runner.new("echo 'hello world'")
       end
 
       include RunnerInterfaceTest
+
+      def test_run
+        out, _ = capture_subprocess_io { @subject.run('file_path.rb') }
+
+        assert_match "hello world", out
+
+        out, _ = capture_subprocess_io { @subject.run }
+
+        assert_match "hello world", out
+      end
 
       def test_matching
         refute @subject.matching?
